+++ src/integration-test/java/org/openlmis/referencedata/web/SupplyLineControllerIntegrationTest.java
@@ -15,9 +15,10 @@
 
 package org.openlmis.referencedata.web;
 
+import static java.util.Collections.emptySet;
+import static java.util.Collections.singleton;
+import static java.util.Collections.singletonList;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasItems;
-import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertEquals;
@@ -30,9 +31,7 @@ import static org.openlmis.referencedata.domain.RightName.SUPPLY_LINES_MANAGE;
 
 import com.jayway.restassured.response.ValidatableResponse;
 import guru.nidi.ramltester.junit.RamlMatchers;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import org.hamcrest.Matchers;
@@ -43,12 +42,11 @@ import org.openlmis.referencedata.domain.RightName;
 import org.openlmis.referencedata.domain.SupplyLine;
 import org.openlmis.referencedata.dto.SupplyLineDto;
 import org.openlmis.referencedata.exception.UnauthorizedException;
-import org.openlmis.referencedata.exception.ValidationMessageException;
 import org.openlmis.referencedata.testbuilder.SupplyLineDataBuilder;
 import org.openlmis.referencedata.util.Message;
 import org.openlmis.referencedata.util.Pagination;
-import org.openlmis.referencedata.util.messagekeys.SupplyLineMessageKeys;
 import org.openlmis.referencedata.utils.AuditLogHelper;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -56,17 +54,17 @@ import org.springframework.http.MediaType;
 @SuppressWarnings("PMD.TooManyMethods")
 public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest {
 
-  private static final String PROGRAM = "program";
-  private static final String SUPERVISORY_NODE = "supervisoryNode";
-  private static final String SUPPLYING_FACILITY = "supplyingFacility";
+  private static final String PROGRAM_ID = "programId";
+  private static final String SUPERVISORY_NODE_ID = "supervisoryNodeId";
+  private static final String SUPPLYING_FACILITY_ID = "supplyingFacilityId";
 
   private static final String RESOURCE_URL = "/api/supplyLines";
-  private static final String SEARCH_URL = RESOURCE_URL + "/search";
   private static final String ID_URL = RESOURCE_URL + "/{id}";
 
   private SupplyLine supplyLine;
   private SupplyLineDto supplyLineDto;
   private UUID supplyLineId;
+  private Pageable pageable;
 
   @Override
   @Before
@@ -77,21 +75,21 @@ public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest 
     supplyLineDto = new SupplyLineDto();
     supplyLine.export(supplyLineDto);
     supplyLineId = supplyLine.getId();
+    pageable = new PageRequest(0, 10);
 
     given(supplyLineRepository.save(any(SupplyLine.class))).willAnswer(new SaveAnswer<>());
   }
 
   @Test
   public void searchShouldReturnUnauthorizedWithoutAuthorization() {
-
-    Map<String, Object> requestBody = getSearchBody();
+    Map<String, Object> requestBody = getSearchParameters();
 
     restAssured
         .given()
-        .body(requestBody)
         .contentType(MediaType.APPLICATION_JSON_VALUE)
+        .params(requestBody)
         .when()
-        .post(SEARCH_URL)
+        .get(RESOURCE_URL)
         .then()
         .statusCode(401);
 
@@ -100,20 +98,16 @@ public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest 
 
   @Test
   public void searchShouldReturnBadRequestOnException() {
-    Map<String, Object> requestBody = getSearchBody();
-
-    doThrow(new ValidationMessageException(
-        SupplyLineMessageKeys.ERROR_SEARCH_LACKS_PARAMS))
-        .when(supplyLineService)
-        .searchSupplyLines(any(Map.class), any(Pageable.class));
+    Map<String, Object> requestBody = getSearchParameters();
 
     restAssured
         .given()
         .header(HttpHeaders.AUTHORIZATION, getTokenHeader())
-        .body(requestBody)
+        .params(requestBody)
+        .param("some-unknown-parameter", "some-value")
         .contentType(MediaType.APPLICATION_JSON_VALUE)
         .when()
-        .post(SEARCH_URL)
+        .get(RESOURCE_URL)
         .then()
         .statusCode(400);
 
@@ -121,27 +115,50 @@ public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest 
   }
 
   @Test
-  public void shouldFindSupplyLines() {
-
-    given(supplyLineService.searchSupplyLines(any(Map.class), any(Pageable.class)))
-        .willReturn(Pagination.getPage(Arrays.asList(supplyLine), null, 1));
-
-    Map<String, Object> requestBody = getSearchBody();
+  public void shouldFindSupplyLinesWithParameters() {
+    given(supplyLineRepository.search(null, null, emptySet(), pageable))
+        .willReturn(Pagination.getPage(singletonList(supplyLine), pageable, 1));
 
     PageImplRepresentation response = restAssured
         .given()
         .header(HttpHeaders.AUTHORIZATION, getTokenHeader())
-        .body(requestBody)
+        .params("page", 1)
+        .params("size", 10)
         .contentType(MediaType.APPLICATION_JSON_VALUE)
         .when()
-        .post(SEARCH_URL)
+        .get(RESOURCE_URL)
         .then()
         .statusCode(200)
         .extract().as(PageImplRepresentation.class);
 
     assertThat(RAML_ASSERT_MESSAGE, restAssured.getLastReport(), RamlMatchers.hasNoViolations());
     assertEquals(1, response.getContent().size());
+  }
+
+  @Test
+  public void shouldFindSupplyLines() {
+    given(supplyLineRepository.search(
+        supplyLine.getProgram().getId(),
+        supplyLine.getSupervisoryNode().getId(),
+        singleton(supplyLine.getSupplyingFacility().getId()),
+        pageable))
+        .willReturn(Pagination.getPage(singletonList(supplyLine), pageable, 1));
+
+    Map<String, Object> requestBody = getSearchParameters();
+
+    PageImplRepresentation response = restAssured
+        .given()
+        .header(HttpHeaders.AUTHORIZATION, getTokenHeader())
+        .params(requestBody)
+        .contentType(MediaType.APPLICATION_JSON_VALUE)
+        .when()
+        .get(RESOURCE_URL)
+        .then()
+        .statusCode(200)
+        .extract().as(PageImplRepresentation.class);
+
     assertThat(RAML_ASSERT_MESSAGE, restAssured.getLastReport(), RamlMatchers.hasNoViolations());
+    assertEquals(1, response.getContent().size());
   }
 
   @Test
@@ -316,41 +333,6 @@ public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest 
   }
 
   @Test
-  public void shouldGetAllSupplyLines() {
-
-    SupplyLine newSupplyLine = generateSupplyLine();
-    List<SupplyLine> storedSupplyLines = Arrays.asList(supplyLine, newSupplyLine);
-    given(supplyLineRepository.findAll()).willReturn(storedSupplyLines);
-
-    restAssured
-        .given()
-        .header(HttpHeaders.AUTHORIZATION, getTokenHeader())
-        .contentType(MediaType.APPLICATION_JSON_VALUE)
-        .when()
-        .get(RESOURCE_URL)
-        .then()
-        .statusCode(200)
-        .body(ID, hasSize(2))
-        .body(ID, hasItems(newSupplyLine.getId().toString(), supplyLine.getId().toString()));
-
-    assertThat(RAML_ASSERT_MESSAGE, restAssured.getLastReport(), RamlMatchers.hasNoViolations());
-  }
-
-  @Test
-  public void getAllShouldReturnUnauthorizedWithoutAuthorization() {
-
-    restAssured
-        .given()
-        .contentType(MediaType.APPLICATION_JSON_VALUE)
-        .when()
-        .get(RESOURCE_URL)
-        .then()
-        .statusCode(401);
-
-    assertThat(RAML_ASSERT_MESSAGE, restAssured.getLastReport(), RamlMatchers.hasNoViolations());
-  }
-
-  @Test
   public void shouldGetSupplyLine() {
 
     given(supplyLineRepository.findOne(supplyLineId)).willReturn(supplyLine);
@@ -453,13 +435,13 @@ public class SupplyLineControllerIntegrationTest extends BaseWebIntegrationTest 
     return new SupplyLineDataBuilder().build();
   }
 
-  private Map<String,Object> getSearchBody() {
-    Map<String, Object> requestBody = new HashMap<>();
-    requestBody.put("page", 1);
-    requestBody.put("size", 10);
-    requestBody.put(PROGRAM, "program-code");
-    requestBody.put(SUPERVISORY_NODE, "node-code");
-    requestBody.put(SUPPLYING_FACILITY, "facility-code");
-    return requestBody;
+  private Map<String,Object> getSearchParameters() {
+    Map<String, Object> paramters = new HashMap<>();
+    paramters.put("page", 1);
+    paramters.put("size", 10);
+    paramters.put(PROGRAM_ID, UUID.randomUUID());
+    paramters.put(SUPERVISORY_NODE_ID, UUID.randomUUID());
+    paramters.put(SUPPLYING_FACILITY_ID, UUID.randomUUID());
+    return paramters;
   }
 }